Remove MDB2 usage.
Currently, the only use of the database has been spun off into a seperate project.
It was originally included to facilitate log recording, however that is now out
of scope. Most likely, anything that does need to be persisted across Ligrev
instances can be stored in a dotfile as a simple array, or something.

If we ever need to use a 'real' database in the future, MDB2 should be
avoided as it is outdated as all hell